House-Inspection_agent/requirements.txt
crewai
python-dotenv
crewai-tools